+++ src/app/(authed)/(dialog)/shared-chats/actions.ts
@@ -1,7 +1,11 @@
 'use server';
 
 import { db } from '@/db';
-import { SharedSchoolConversationFileMapping, sharedSchoolConversationTable } from '@/db/schema';
+import {
+  SharedSchoolConversationFileMapping,
+  SharedSchoolConversationInsertModel,
+  sharedSchoolConversationTable,
+} from '@/db/schema';
 import { getUser } from '@/auth/utils';
 import { dbDeleteSharedSchoolChatByIdAndUserId } from '@/db/functions/shared-school-chat';
 import { DEFAULT_CHAT_MODEL } from '@/app/api/chat/models';
@@ -16,6 +20,33 @@ export async function dbDeleteSharedChatAction({ id }: { id: string }) {
   });
 
   return deletedSharedChat;
+}
+
+export async function createNewSharedSchoolChatAction(
+  data: Omit<SharedSchoolConversationInsertModel, 'userId'>,
+) {
+  const user = await getUser();
+
+  if (user.school === undefined) {
+    throw Error('User is not part of a school');
+  }
+
+  if (user.school.userRole !== 'teacher') {
+    throw Error('Only a teacher can create shared chats');
+  }
+
+  const insertedSharedChat = (
+    await db
+      .insert(sharedSchoolConversationTable)
+      .values({ ...data, userId: user.id })
+      .returning()
+  )[0];
+
+  if (insertedSharedChat === undefined) {
+    throw Error('Could not insert chat');
+  }
+
+  return insertedSharedChat;
 }
 
 export async function dbCreateSharedSchoolChat({ userId }: { userId: string }) {
